+++ DataProcessing/DataProcessing.csproj
@@ -3,21 +3,17 @@
 	<PropertyGroup>
 		<Configuration Condition=" '$(Configuration)' == '' ">Release</Configuration>
 		<Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-		<TargetFramework>net6.0</TargetFramework>
-		<Product>QuantConnect.CoinAPI.Converter</Product>
-		<AssemblyName>QuantConnect.CoinAPI.Converter</AssemblyName>
-		<RootNamespace>QuantConnect.CoinAPI.Converter</RootNamespace>
-		<AssemblyTitle>QuantConnect.CoinAPI.Converter</AssemblyTitle>
 		<OutputType>Exe</OutputType>
+		<TargetFramework>net6.0</TargetFramework>
+		<AssemblyName>process</AssemblyName>
+		<CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
 		<OutputPath>bin\$(Configuration)</OutputPath>
 		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-		<CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
-		<GenerateAssemblyInfo>false</GenerateAssemblyInfo>
 		<Description>QuantConnect LEAN CoinAPI Converter Data Source: CoinAPI Converter Data Source plugin for Lean</Description>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
-
+	
 	<PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
 		<DebugType>full</DebugType>
 		<OutputPath>bin\Debug\</OutputPath>
